Update reward validation and (bonus) selection mechanism
Reward functions are now members of a table, keyed to each reward's hash value, for easy selection by the scripts and to eliminate multiple if-then-else checks to identify valid rewards.

Hash values for each reward type and subtype have been added to the appropriate ingame DB tables to facilitate reward-to-hash referencing; new tables to facilitate hash-to-reward referencing were previously added.